requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
iniconfig==1.1.1
Markdown==3.3.4
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
